article-sumariser-agent/requirements.txt
requests
python-dotenv
newspaper3k
openai
transformers
torch
sentencepiece
gTTS
tqdm
apscheduler
